refactor(retargeters): move the analog gripper retargeter out of SO101
The node reads one controller's analog trigger and emits a unitless closedness
in [0, 1]; nothing in it knows about jaw travel, so it does not belong in the
SO101 subpackage. Move only -- the class keeps its name here, and the rename
follows in the next commit.

SO101ClutchRetargeter stays put: its clutch really is about a 5-DOF arm.
